backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.35.0",
]
